requirements.txt
Pillow>=9.0
numpy>=1.21
opencv-python>=4.5
fastapi>=0.100
uvicorn>=0.23
python-multipart>=0.0.6

